# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Delphi API",
    "description": "Compare organizational documents with traceable evidence and human review.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/analyses/{analysis_id}/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Start Run",
        "operationId": "start_run",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRun"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/runs/{run_id}/functions": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Functions",
        "operationId": "list_functions",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunctionResponse"
                  },
                  "title": "Response List Functions"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorResponse"
      },
...
      "FunctionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "side": {
            "type": "string",
            "enum": [
              "before",
              "after"
            ],
            "title": "Side"
          },
          "owner_unit_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Owner Unit Ids"
          },
          "actor_original": {
            "type": "string",
            "title": "Actor Original"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "object": {
            "type": "string",
            "title": "Object"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "condition": {
            "type": "string",
            "title": "Condition"
          },
          "modality": {
            "type": "string",
            "title": "Modality"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "side",
          "owner_unit_ids",
          "actor_original",
          "action",
          "object",
          "scope",
          "condition",
          "modality",
          "source_ids"
        ],
        "title": "FunctionResponse"
      },
...
      "RunAccepted": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "interrupted"
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "state"
        ],
        "title": "RunAccepted"
      },
      "RunCoverage": {
        "properties": {
          "allow_partial": {
            "type": "boolean",
            "title": "Allow Partial"
          },
          "input_partial": {
            "type": "boolean",
            "title": "Input Partial"
          },
          "input_warnings": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "type": "object",
            "title": "Input Warnings"
          },
          "total_documents": {
            "type": "integer",
            "title": "Total Documents"
          },
          "total_sources": {
            "type": "integer",
            "title": "Total Sources"
          },
          "processed_sources": {
            "type": "integer",
            "title": "Processed Sources"
          },
          "before_functions": {
            "type": "integer",
            "title": "Before Functions"
          },
          "compared_before_functions": {
            "type": "integer",
            "title": "Compared Before Functions"
          },
          "after_functions": {
            "type": "integer",
            "title": "After Functions"
          },
          "reviewed_after_functions": {
            "type": "integer",
            "title": "Reviewed After Functions"
          },
          "structure_units": {
            "type": "integer",
            "title": "Structure Units"
          },
          "reviewed_structure_units": {
            "type": "integer",
            "title": "Reviewed Structure Units"
          },
          "unprocessed_source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Unprocessed Source Ids"
          },
          "unreviewed_function_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Unreviewed Function Ids"
          }
        },
        "type": "object",
        "required": [
          "allow_partial",
          "input_partial",
          "input_warnings",
          "total_documents",
          "total_sources",
          "processed_sources",
          "before_functions",
          "compared_before_functions",
          "after_functions",
          "reviewed_after_functions",
          "structure_units",
          "reviewed_structure_units",
          "unprocessed_source_ids",
          "unreviewed_function_ids"
        ],
        "title": "RunCoverage"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "interrupted"
            ],
            "title": "State"
          },
          "stage": {
            "type": "string",
            "enum": [
              "queued",
              "extracting",
              "comparing",
              "checking_risks",
              "validating",
              "completed",
              "partial",
              "failed",
              "interrupted"
            ],
            "title": "Stage"
          },
          "output_language": {
            "type": "string",
            "enum": [
              "ru",
              "kk",
              "en"
            ],
            "title": "Output Language"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version"
          },
          "review_revision": {
            "type": "integer",
            "title": "Review Revision"
          },
          "coverage": {
            "$ref": "#/components/schemas/RunCoverage"
          },
          "structure": {
            "items": {
              "$ref": "#/components/schemas/StructureChange"
            },
            "type": "array",
            "title": "Structure"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "units": {
            "items": {
              "$ref": "#/components/schemas/UnitResponse"
            },
            "type": "array",
            "title": "Units"
          },
          "finding_count": {
            "type": "integer",
            "title": "Finding Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "analysis_id",
          "state",
          "stage",
          "output_language",
          "model",
          "pipeline_version",
          "review_revision",
          "coverage",
          "structure",
          "errors",
          "created_at",
          "started_at",
          "finished_at",
          "units",
          "finding_count"
        ],
        "title": "RunDetail"
      },
...
      "StartRun": {
        "properties": {
          "output_language": {
            "type": "string",
            "enum": [
              "ru",
              "kk",
              "en"
            ],
            "title": "Output Language"
          },
          "allow_partial": {
            "type": "boolean",
            "title": "Allow Partial"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "output_language",
          "allow_partial"
        ],
        "title": "StartRun"
      },
      "StructureChange": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "before_unit_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Before Unit Ids"
          },
          "after_unit_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "After Unit Ids"
          },
          "status": {
            "type": "string",
            "enum": [
              "retained",
              "newly_listed",
              "transformed",
              "unmatched"
            ],
            "title": "Status"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "id",
          "before_unit_ids",
          "after_unit_ids",
          "status",
          "source_ids",
          "explanation"
        ],
        "title": "StructureChange"
      },
...
      "UnitResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "side": {
            "type": "string",
            "enum": [
              "before",
              "after"
            ],
            "title": "Side"
          },
          "kind": {
            "type": "string",
            "enum": [
              "department",
              "role",
              "group"
            ],
            "title": "Kind"
          },
          "name_original": {
            "type": "string",
            "title": "Name Original"
          },
          "parent_unit_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Unit Id"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Source Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "side",
          "kind",
          "name_original",
          "parent_unit_id",
          "source_ids"
        ],
        "title": "UnitResponse"
      },
...
    }
  }
}